+++ sdk/ai/azopenai/client.go
@@ -29,10 +29,65 @@ type Client struct {
 	clientData
 }
 
+// AddUploadPart - Adds a Part to an Upload object. A Part represents a chunk of bytes from the file you are trying to upload.
+// Each Part can be at most 64 MB, and you can add Parts until you hit the Upload maximum of 8 GB.
+// It is possible to add multiple Parts in parallel. You can decide the intended order of the Parts when you complete the
+// Upload.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - uploadID - The ID of the upload associated with this operation.
+//   - data - The chunk of bytes for this Part.
+//   - options - AddUploadPartOptions contains the optional parameters for the Client.AddUploadPart method.
+func (client *Client) AddUploadPart(ctx context.Context, uploadID string, data io.ReadSeekCloser, options *AddUploadPartOptions) (AddUploadPartResponse, error) {
+	var err error
+	req, err := client.addUploadPartCreateRequest(ctx, uploadID, data, options)
+	if err != nil {
+		return AddUploadPartResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return AddUploadPartResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
+		err = client.newError(httpResp)
+		return AddUploadPartResponse{}, err
+	}
+	resp, err := client.addUploadPartHandleResponse(httpResp)
+	return resp, err
+}
+
+// addUploadPartCreateRequest creates the AddUploadPart request.
+func (client *Client) addUploadPartCreateRequest(ctx context.Context, uploadID string, data io.ReadSeekCloser, options *AddUploadPartOptions) (*policy.Request, error) {
+	urlPath := "/uploads/{upload_id}/parts"
+	if uploadID == "" {
+		return nil, errors.New("parameter uploadID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{upload_id}", url.PathEscape(uploadID))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, client.formatURL(urlPath, nil))
+	if err != nil {
+		return nil, err
+	}
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := setMultipartFormData(req, data, NoFilenameChange{}); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// addUploadPartHandleResponse handles the AddUploadPart response.
+func (client *Client) addUploadPartHandleResponse(resp *http.Response) (AddUploadPartResponse, error) {
+	result := AddUploadPartResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.UploadPart); err != nil {
+		return AddUploadPartResponse{}, err
+	}
+	return result, nil
+}
+
 // CancelBatch - Gets details for a single batch specified by the given batchID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - batchID - The identifier of the batch.
 //   - options - CancelBatchOptions contains the optional parameters for the Client.CancelBatch method.
 func (client *Client) CancelBatch(ctx context.Context, batchID string, options *CancelBatchOptions) (CancelBatchResponse, error) {
@@ -77,11 +132,115 @@ func (client *Client) cancelBatchHandleResponse(resp *http.Response) (CancelBatc
 	return result, nil
 }
 
+// CancelUpload - Cancels the Upload. No Parts may be added after an Upload is cancelled.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - uploadID - The ID of the upload associated with this operation.
+//   - options - CancelUploadOptions contains the optional parameters for the Client.CancelUpload method.
+func (client *Client) CancelUpload(ctx context.Context, uploadID string, options *CancelUploadOptions) (CancelUploadResponse, error) {
+	var err error
+	req, err := client.cancelUploadCreateRequest(ctx, uploadID, options)
+	if err != nil {
+		return CancelUploadResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return CancelUploadResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
+		err = client.newError(httpResp)
+		return CancelUploadResponse{}, err
+	}
+	resp, err := client.cancelUploadHandleResponse(httpResp)
+	return resp, err
+}
+
+// cancelUploadCreateRequest creates the CancelUpload request.
+func (client *Client) cancelUploadCreateRequest(ctx context.Context, uploadID string, options *CancelUploadOptions) (*policy.Request, error) {
+	urlPath := "/uploads/{upload_id}/cancel"
+	if uploadID == "" {
+		return nil, errors.New("parameter uploadID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{upload_id}", url.PathEscape(uploadID))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, client.formatURL(urlPath, nil))
+	if err != nil {
+		return nil, err
+	}
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// cancelUploadHandleResponse handles the CancelUpload response.
+func (client *Client) cancelUploadHandleResponse(resp *http.Response) (CancelUploadResponse, error) {
+	result := CancelUploadResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Upload); err != nil {
+		return CancelUploadResponse{}, err
+	}
+	return result, nil
+}
+
+// CompleteUpload - Completes the Upload.
+// Within the returned Upload object, there is a nested File object that is ready to use in the rest of the platform.
+// You can specify the order of the Parts by passing in an ordered list of the Part IDs.
+// The number of bytes uploaded upon completion must match the number of bytes initially specified when creating the Upload
+// object. No Parts may be added after an Upload is completed.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - uploadID - The ID of the upload associated with this operation.
+//   - requestBody - The request body for the completion operation.
+//   - options - CompleteUploadOptions contains the optional parameters for the Client.CompleteUpload method.
+func (client *Client) CompleteUpload(ctx context.Context, uploadID string, requestBody CompleteUploadRequest, options *CompleteUploadOptions) (CompleteUploadResponse, error) {
+	var err error
+	req, err := client.completeUploadCreateRequest(ctx, uploadID, requestBody, options)
+	if err != nil {
+		return CompleteUploadResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return CompleteUploadResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
+		err = client.newError(httpResp)
+		return CompleteUploadResponse{}, err
+	}
+	resp, err := client.completeUploadHandleResponse(httpResp)
+	return resp, err
+}
+
+// completeUploadCreateRequest creates the CompleteUpload request.
+func (client *Client) completeUploadCreateRequest(ctx context.Context, uploadID string, requestBody CompleteUploadRequest, options *CompleteUploadOptions) (*policy.Request, error) {
+	urlPath := "/uploads/{upload_id}/complete"
+	if uploadID == "" {
+		return nil, errors.New("parameter uploadID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{upload_id}", url.PathEscape(uploadID))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, client.formatURL(urlPath, nil))
+	if err != nil {
+		return nil, err
+	}
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, requestBody); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// completeUploadHandleResponse handles the CompleteUpload response.
+func (client *Client) completeUploadHandleResponse(resp *http.Response) (CompleteUploadResponse, error) {
+	result := CompleteUploadResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Upload); err != nil {
+		return CompleteUploadResponse{}, err
+	}
+	return result, nil
+}
+
 // CreateBatch - Creates and executes a batch from an uploaded file of requests. Response includes details of the enqueued
 // job including job status. The ID of the result file is added to the response once complete.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - createBatchRequest - The specification of the batch to create and execute.
 //   - options - CreateBatchOptions contains the optional parameters for the Client.CreateBatch method.
 func (client *Client) CreateBatch(ctx context.Context, createBatchRequest BatchCreateRequest, options *CreateBatchOptions) (CreateBatchResponse, error) {
@@ -125,10 +284,63 @@ func (client *Client) createBatchHandleResponse(resp *http.Response) (CreateBatc
 	return result, nil
 }
 
+// CreateUpload - Creates an intermediate Upload object that you can add Parts to. Currently, an Upload can accept at most
+// 8 GB in total and expires after an hour after you create it.
+// Once you complete the Upload, we will create a File object that contains all the parts you uploaded. This File is usable
+// in the rest of our platform as a regular File object.
+// For certain purposes, the correct mime_type must be specified. Please refer to documentation for the supported MIME types
+// for your use case.
+// For guidance on the proper filename extensions for each purpose, please follow the documentation on creating a File.
+// If the operation fails it returns an *azcore.ResponseError type.
+//
+// Generated from API version 2024-08-01-preview
+//   - requestBody - The request body for the operation options.
+//   - options - CreateUploadOptions contains the optional parameters for the Client.CreateUpload method.
+func (client *Client) CreateUpload(ctx context.Context, requestBody CreateUploadRequest, options *CreateUploadOptions) (CreateUploadResponse, error) {
+	var err error
+	req, err := client.createUploadCreateRequest(ctx, requestBody, options)
+	if err != nil {
+		return CreateUploadResponse{}, err
+	}
+	httpResp, err := client.internal.Pipeline().Do(req)
+	if err != nil {
+		return CreateUploadResponse{}, err
+	}
+	if !runtime.HasStatusCode(httpResp, http.StatusOK) {
+		err = client.newError(httpResp)
+		return CreateUploadResponse{}, err
+	}
+	resp, err := client.createUploadHandleResponse(httpResp)
+	return resp, err
+}
+
+// createUploadCreateRequest creates the CreateUpload request.
+func (client *Client) createUploadCreateRequest(ctx context.Context, requestBody CreateUploadRequest, options *CreateUploadOptions) (*policy.Request, error) {
+	urlPath := "/uploads"
+	req, err := runtime.NewRequest(ctx, http.MethodPost, client.formatURL(urlPath, nil))
+	if err != nil {
+		return nil, err
+	}
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	if err := runtime.MarshalAsJSON(req, requestBody); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+// createUploadHandleResponse handles the CreateUpload response.
+func (client *Client) createUploadHandleResponse(resp *http.Response) (CreateUploadResponse, error) {
+	result := CreateUploadResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Upload); err != nil {
+		return CreateUploadResponse{}, err
+	}
+	return result, nil
+}
+
 // DeleteFile - Delete a previously uploaded file.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - fileID - The ID of the file to delete.
 //   - options - DeleteFileOptions contains the optional parameters for the Client.DeleteFile method.
 func (client *Client) DeleteFile(ctx context.Context, fileID string, options *DeleteFileOptions) (DeleteFileResponse, error) {
@@ -176,7 +388,7 @@ func (client *Client) deleteFileHandleResponse(resp *http.Response) (DeleteFileR
 // GenerateSpeechFromText - Generates text-to-speech audio from the input text.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - body - A representation of the request options that control the behavior of a text-to-speech operation.
 //   - options - GenerateSpeechFromTextOptions contains the optional parameters for the Client.GenerateSpeechFromText method.
 func (client *Client) GenerateSpeechFromText(ctx context.Context, body SpeechGenerationOptions, options *GenerateSpeechFromTextOptions) (GenerateSpeechFromTextResponse, error) {
@@ -204,7 +416,7 @@ func (client *Client) generateSpeechFromTextCreateRequest(ctx context.Context, b
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	runtime.SkipBodyDownload(req)
 	req.Raw().Header["Accept"] = []string{"application/octet-stream, application/json"}
@@ -218,7 +430,7 @@ func (client *Client) generateSpeechFromTextCreateRequest(ctx context.Context, b
 // be transcribed in the written language corresponding to the language it was spoken in.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - deploymentID - Specifies either the model deployment name (when using Azure OpenAI) or model name (when using non-Azure
 //     OpenAI) to use for this request.
 //   - file - The audio data to transcribe. This must be the binary content of a file in one of the supported media formats: flac,
@@ -251,7 +463,7 @@ func (client *Client) getAudioTranscriptionInternalCreateRequest(ctx context.Con
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := setMultipartFormData(req, file, *body); err != nil {
@@ -273,7 +485,7 @@ func (client *Client) getAudioTranscriptionInternalHandleResponse(resp *http.Res
 // data.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - deploymentID - Specifies either the model deployment name (when using Azure OpenAI) or model name (when using non-Azure
 //     OpenAI) to use for this request.
 //   - file - The audio data to translate. This must be the binary content of a file in one of the supported media formats: flac,
@@ -306,7 +518,7 @@ func (client *Client) getAudioTranslationInternalCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := setMultipartFormData(req, file, *body); err != nil {
@@ -327,7 +539,7 @@ func (client *Client) getAudioTranslationInternalHandleResponse(resp *http.Respo
 // GetBatch - Gets details for a single batch specified by the given batchID.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - batchID - The identifier of the batch.
 //   - options - GetBatchOptions contains the optional parameters for the Client.GetBatch method.
 func (client *Client) GetBatch(ctx context.Context, batchID string, options *GetBatchOptions) (GetBatchResponse, error) {
@@ -376,7 +588,7 @@ func (client *Client) getBatchHandleResponse(resp *http.Response) (GetBatchRespo
 // and generate text that continues from or "completes" provided prompt data.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - body - The configuration information for a chat completions request. Completions support a wide variety of tasks and generate
 //     text that continues from or "completes" provided prompt data.
 //   - options - GetChatCompletionsOptions contains the optional parameters for the Client.GetChatCompletions method.
@@ -406,7 +618,7 @@ func (client *Client) getChatCompletionsCreateRequest(ctx context.Context, body 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
@@ -428,7 +640,7 @@ func (client *Client) getChatCompletionsHandleResponse(resp *http.Response) (Get
 // text that continues from or "completes" provided prompt data.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - body - The configuration information for a completions request. Completions support a wide variety of tasks and generate
 //     text that continues from or "completes" provided prompt data.
 //   - options - GetCompletionsOptions contains the optional parameters for the Client.GetCompletions method.
@@ -458,7 +670,7 @@ func (client *Client) getCompletionsCreateRequest(ctx context.Context, body Comp
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
@@ -479,7 +691,7 @@ func (client *Client) getCompletionsHandleResponse(resp *http.Response) (GetComp
 // GetEmbeddings - Return the embeddings for a given prompt.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - body - The configuration information for an embeddings request. Embeddings measure the relatedness of text strings and
 //     are commonly used for search, clustering, recommendations, and other similar scenarios.
 //   - options - GetEmbeddingsOptions contains the optional parameters for the Client.GetEmbeddings method.
@@ -509,7 +721,7 @@ func (client *Client) getEmbeddingsCreateRequest(ctx context.Context, body Embed
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
@@ -530,7 +742,7 @@ func (client *Client) getEmbeddingsHandleResponse(resp *http.Response) (GetEmbed
 // GetFile - Returns information about a specific file. Does not retrieve file content.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - fileID - The ID of the file to retrieve.
 //   - options - GetFileOptions contains the optional parameters for the Client.GetFile method.
 func (client *Client) GetFile(ctx context.Context, fileID string, options *GetFileOptions) (GetFileResponse, error) {
@@ -578,7 +790,7 @@ func (client *Client) getFileHandleResponse(resp *http.Response) (GetFileRespons
 // GetFileContent - Returns information about a specific file. Does not retrieve file content.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - fileID - The ID of the file to retrieve.
 //   - options - GetFileContentOptions contains the optional parameters for the Client.GetFileContent method.
 func (client *Client) GetFileContent(ctx context.Context, fileID string, options *GetFileContentOptions) (GetFileContentResponse, error) {
@@ -626,7 +838,7 @@ func (client *Client) getFileContentHandleResponse(resp *http.Response) (GetFile
 // GetImageGenerations - Creates an image given a prompt.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - body - Represents the request data used to generate images.
 //   - options - GetImageGenerationsOptions contains the optional parameters for the Client.GetImageGenerations method.
 func (client *Client) GetImageGenerations(ctx context.Context, body ImageGenerationOptions, options *GetImageGenerationsOptions) (GetImageGenerationsResponse, error) {
@@ -655,7 +867,7 @@ func (client *Client) getImageGenerationsCreateRequest(ctx context.Context, body
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2024-07-01-preview")
+	reqQP.Set("api-version", "2024-08-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if err := runtime.MarshalAsJSON(req, body); err != nil {
@@ -676,7 +888,7 @@ func (client *Client) getImageGenerationsHandleResponse(resp *http.Response) (Ge
 // listBatches - Gets a list of all batches owned by the Azure OpenAI resource.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - options - ListBatchesOptions contains the optional parameters for the Client.listBatches method.
 func (client *Client) listBatches(ctx context.Context, options *ListBatchesOptions) (ListBatchesResponse, error) {
 	var err error
@@ -727,7 +939,7 @@ func (client *Client) listBatchesHandleResponse(resp *http.Response) (ListBatche
 // ListFiles - Gets a list of previously uploaded files.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - options - ListFilesOptions contains the optional parameters for the Client.ListFiles method.
 func (client *Client) ListFiles(ctx context.Context, options *ListFilesOptions) (ListFilesResponse, error) {
 	var err error
@@ -775,7 +987,7 @@ func (client *Client) listFilesHandleResponse(resp *http.Response) (ListFilesRes
 // UploadFile - Uploads a file for use by other operations.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2024-07-01-preview
+// Generated from API version 2024-08-01-preview
 //   - file - The file data (not filename) to upload.
 //   - purpose - The intended purpose of the file.
 //   - options - UploadFileOptions contains the optional parameters for the Client.UploadFile method.